requirements.txt
# MSc project — notebook + EDA (pin for reproducibility)
jupyter>=1.0.0,<2
ipykernel>=6.29.0,<7
pandas>=2.1.0,<3
numpy>=1.26.0,<3
matplotlib>=3.8.0,<4
seaborn>=0.13.0,<0.14
requests>=2.28.0,<3
Pillow>=10.0.0,<12
gdown>=4.7.0,<6
ImageHash>=4.3.0,<5
